requirements.txt
diffusers
transformers
accelerate
torch
safetensors
numpy
matplotlib
pillow